requirements.txt
langchain
ipykernel
python-dotenv

langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia

langchain-text-splitters

# langchain-openai

langchain-ollama

langchain_huggingface
sentence_transformers


faiss-cpu
langchain_chroma
langchain-core
langchain_classic

streamlit

langchain_groq

fastapi
uvicorn
langserve[all]
sse_starlette

langchain_google_genai
langgraph

SQLAlchemy
validators
youtube_transcript_api
pytube
unstructured

langchain_experimental
huggingface_hub

gradio